move filters routes
